Split Parser into Parser and (internal)ArgumentParser
 - Parser prepare the rules and
 - ArgumentParser parses the argument into the ParseRules
 - Parser writes the ParseRule into the options object.
