+++ CommandLineParser/CommandLineParser.csproj
@@ -42,6 +42,7 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="Attribute\OptionAttribute.cs" />
+    <Compile Include="Parser\ArgumentParser.cs" />
     <Compile Include="Parser\Parser.cs" />
     <Compile Include="Parser\ParseRule.cs" />
     <Compile Include="Parser\ParseState.cs" />
